+++ web/documentserver-example/php-laravel/app/Http/Controllers/EditorController.php
@@ -255,7 +255,8 @@ class EditorController extends Controller
                 // prevent switch the document from the viewing into the editing mode for anonymous users
                 config.events['onRequestEditRights'] = onRequestEditRights;
                 config.events['onRequestHistory'] = onRequestHistory;
-                config.events['onRequestHistoryData'] = onRequestHistoryData;";
+                config.events['onRequestHistoryData'] = onRequestHistoryData;
+                config.events['onRequestReferenceSource'] = onRequestReferenceSource;";
             if ($user['id'] != 'uid-3') {
                 $historyLayout .= "config.events['onRequestHistoryClose'] = onRequestHistoryClose;
                 config.events['onRequestRestore'] = onRequestRestore;";