+++ web/documentserver-example/php-laravel/resources/views/editor.blade.php
@@ -157,6 +157,53 @@
             }
         };
 
+        var onRequestReferenceSource = function (event) {
+          innerAlert("onRequestReferenceSource");
+          let xhr = new XMLHttpRequest();
+          xhr.open("GET", "/files");
+          xhr.setRequestHeader("Content-Type", "application/json");
+          xhr.send();
+          xhr.onload = function () {
+            if (xhr.status === 200) {
+              innerAlert(JSON.parse(xhr.responseText));
+              let fileList = JSON.parse(xhr.responseText);
+              let firstXlsxName;
+              let file;
+              for (file of fileList) {
+                if (file["title"]) {
+                  if (getFileExt(file["title"]) === "xlsx")
+                  {
+                    firstXlsxName = file["title"];
+                    break;
+                  }
+                }
+              }
+              if (firstXlsxName) {
+                let data = {
+                  directUrl : !!config.document.directUrl,
+                  path : firstXlsxName
+                };
+                let xhr = new XMLHttpRequest();
+                xhr.open("POST", "/api/files/reference");
+                xhr.setRequestHeader("Content-Type", "application/json");
+                xhr.setRequestHeader('X-CSRF-TOKEN', "{{ csrf_token() }}")
+                xhr.send(JSON.stringify(data));
+                xhr.onload = function () {
+                  if (xhr.status === 200) {
+                    docEditor.setReferenceSource(JSON.parse(xhr.responseText));
+                  } else {
+                    innerAlert("/reference - bad status");
+                  }
+                }
+              } else {
+                innerAlert("No *.xlsx files");
+              }
+            } else {
+              innerAlert("/files - bad status");
+            }
+          }
+        };
+
         // the user is trying to get link for opening the document which contains a bookmark,
         // scrolling to the bookmark position
         var onMakeActionLink = function(event) {
@@ -362,6 +409,13 @@
             {!! $history !!}
 
             docEditor = new DocsAPI.DocEditor("iframeEditor", config);
+        };
+
+        const getFileExt = function (fileName) {
+          if (fileName.indexOf(".")) {
+            return fileName.split('.').reverse()[0];
+          }
+          return false;
         };
 
         if (window.addEventListener) {